[HIP-CPU] Strictly ISO conforming overload control
